Update codes to remove use of macros, fix python errors, fix const issues
    commit 3f32efb75d9e7c23742e0ec846f1172353488622
    Author: Michael Jackson <mike.jackson@bluequartz.net>
    Date:   Wed Oct 30 14:51:57 2019 -0400

        Fix bug using std::accumulate due to casting issue.

        Passing in "1" will force the computation to be done in int32_t. When using large
        Tuple dimensions we will over flow.


    commit 5aa467054769cfe02060763eb1c9991fde90397f
    Author: Michael Jackson <mike.jackson@bluequartz.net>
    Date:   Tue Jun 25 08:50:25 2019 -0400

        replace 0 with nullptr when pointers are the type.


    commit 87ba1e0f915f28f59e69572424ec2a30e74267a8
    Author: Michael Jackson <mike.jackson@bluequartz.net>
    Date:   Thu Jun 13 18:18:55 2019 -0400

        Massive code modifications to clean up and reduce macro clutter

        * Remove use of all SIMPL_* macros replacing with actual C++ code
        * Remove the include directive to DataContainerArray in AbstractFilter instead using a forward declaration and alias
        * Add needed include directives or forward declarations to all effected classes

+++ SIMPLVtkLib/SIMPLBridge/SIMPLVtkBridge.cpp
@@ -1,4 +1,10 @@
 /* ============================================================================
+
+#include "SIMPLib/DataContainers/DataContainer.h"
+#include "SIMPLib/DataContainers/DataContainerArray.h"
+
+
+
  * Copyright (c) 2009-2015 BlueQuartz Software, LLC
  *
  * Redistribution and use in source and binary forms, with or without modification,
@@ -429,7 +435,7 @@ bool SIMPLVtkBridge::WrapAttrMatrixData(AttributeMatrix::Pointer attrMat, Wrappe
 
   std::vector<size_t> tupleDims = attrMat->getTupleDimensions();
 
-  size_t numTuples = std::accumulate(tupleDims.begin(), tupleDims.end(), 1, std::multiplies<>());
+  size_t numTuples = std::accumulate(tupleDims.begin(), tupleDims.end(), static_cast<size_t>(1)), std::multiplies<>());
 
   if(tuplesReq == numTuples)
   {
